+++ moncic/__main__.py
@@ -2,9 +2,7 @@ import logging
 import sys
 from typing import Protocol
 
-import moncic
 from moncic import cli, exceptions
-from moncic.utils import argparse
 
 log = logging.getLogger(__name__)
 
@@ -17,29 +15,7 @@ class Handler(Protocol):
 
 
 def main() -> int | None:
-    parser = argparse.ArgumentParser(description="CI tool")
-    parser.add_argument(
-        "--version", action="version", version=moncic.__version__
-    )
-    parser.add_argument(
-        "-v",
-        "--verbose",
-        action="store_true",
-        help="verbose output",
-        shared=True,
-    ),
-    parser.add_argument(
-        "--debug",
-        action="store_true",
-        help="debugging output",
-        shared=True,
-    ),
-    subparsers = parser.add_subparsers(
-        help="sub-command help", dest="handler", required=True
-    )
-
-    for cls in cli.MAIN_COMMANDS:
-        cls.make_subparser(subparsers)
+    parser = cli.make_argparser()
 
     try:
         args = parser.parse_args()