+++ moncic/cli/__init__.py
@@ -1,6 +1,12 @@
 # Import so they can be registered in monci.MAIN_COMMANDS
 from . import build, image, maint, query, run  # noqa
 from .base import FAIL_EXCEPTIONS, SUCCESS_EXCEPTIONS, run_main
-from .moncic import MAIN_COMMANDS
+from .moncic import MAIN_COMMANDS, make_argparser
 
-__all__ = ["run_main", "MAIN_COMMANDS", "FAIL_EXCEPTIONS", "SUCCESS_EXCEPTIONS"]
+__all__ = [
+    "make_argparser",
+    "run_main",
+    "MAIN_COMMANDS",
+    "FAIL_EXCEPTIONS",
+    "SUCCESS_EXCEPTIONS",
+]