+++ moncic/cli/tests/test_moncic.py
@@ -1,0 +1,24 @@
+from pathlib import Path
+from unittest import mock
+
+from moncic.exceptions import Fail
+from moncic.provision.image import ConfiguredImage
+from moncic.unittest import TestCase
+from moncic.cli.moncic import MoncicCommand, make_argparser
+from moncic.moncic import MoncicConfig
+
+
+class CliImageTests(TestCase):
+    def test_imageconfs_from_imagedir(self) -> None:
+        sample_imagedir = Path("/does-not-exist")
+        parser = make_argparser()
+        args = parser.parse_args(
+            ["bootstrap", "--imagedir", sample_imagedir.as_posix(), "testimage"]
+        )
+        cmd = MoncicCommand(args)
+        config = cmd.moncic.config
+        self.assertEqual(config.imagedir, sample_imagedir)
+        self.assertEqual(
+            config.imageconfdirs,
+            [MoncicConfig.xdg_local_config_dir(), sample_imagedir],
+        )