+++ NEWS.md
@@ -5,6 +5,7 @@
 * Fedora 42 and Fedora 44 containers now work with podman
 * Added instructions for running autopkgtests from moncic-ci images
 * Allow bootstrapping a mix of bootstrapped and non-bootstrapped images (#127)
+* Load image configuration also from nspawn imagedirs (#126)
 
 # Version 0.29